requirements.txt
cloudinary
dj-database-url==0.5.0
Django==3.2
django-cloudinary-storage
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
mock==4.0.2
Pillow==7.0.0
proto-plus==1.13.0
protobuf==3.14.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-filter
mixer
pytest-django
django-utils-six
